tests/wpt/tests/html/semantics/popovers/popover-types-with-hints.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popover.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  function getPopovers() {
    return Array.from(document.currentScript.parentElement.querySelectorAll('[popover]'));
  }
  function assertState(expectedState,description) {
    description = description || 'Error';
    const popovers = getPopovers();
    const n = popovers.length;
    assert_equals(expectedState.length,n,'Invalid expected state length');
    for(let i=0;i<n;++i) {
      const html = '<' + popovers[i].outerHTML.split('<')[1] + '</div>';
      assert_equals(popovers[i].matches(':popover-open'),expectedState[i],`${description}, index ${i} (${html})`);
    }
  }
</script>

<div>
  <div popover>Popover</div>
  <div popover=hint>Hint</div>
  <div popover=manual>Async</div>
  <div popover=manual>Async</div>
  <script>
  {
    const auto = getPopovers()[0];
    const hint = getPopovers()[1];
    const manual = getPopovers()[2];
    const manual2 = getPopovers()[3];
    test(() => {
      assertState([false,false,false,false]);
      auto.showPopover();
      assertState([true,false,false,false]);
      hint.showPopover();
      assertState([true,true,false,false]);
      manual.showPopover();
      assertState([true,true,true,false]);
      manual2.showPopover();
      assertState([true,true,true,true]);
      hint.hidePopover();
      assertState([true,false,true,true]);
      auto.hidePopover();
      assertState([false,false,true,true]);
      auto.showPopover();
      hint.showPopover();
      assertState([true,true,true,true]);
      auto.hidePopover();
      assertState([false,true,true,true]);
      hint.hidePopover();
      manual.hidePopover();
      assertState([false,false,false,true]);
      manual2.hidePopover();
      assertState([false,false,false,false]);
    },'manuals do not close popovers');

    test(() => {
      assertState([false,false,false,false]);
      hint.showPopover();
      manual.showPopover();
      manual2.showPopover();
      assertState([false,true,true,true]);
      auto.showPopover();
      assertState([true,false,true,true]);
      auto.hidePopover();
      assertState([false,false,true,true]);
      manual.hidePopover();
      manual2.hidePopover();
      assertState([false,false,false,false]);
    },'autos close hints but not manuals');
  }
  </script>
</div>

<div>
  <div popover=hint>Hint
    <div popover=hint>Nested hint</div>
  </div>
  <script>
  test(() => {
    const hint1 = getPopovers()[0];
    const hint2 = getPopovers()[1];
    hint1.showPopover();
    assertState([true,false]);
    hint2.showPopover();
    assertState([true,true]);
    hint1.hidePopover();
    assertState([false,false]);
  },'You can nest hint popovers');
  </script>
</div>

<div>
  <div popover="hint">Hint
    <div popover>Nested auto (note - never visible, since inside display:none subtree)</div>
  </div>
  <script>
  test(() => {
    const hint = getPopovers()[0];
    const auto = getPopovers()[1];
    hint.showPopover();
    assertState([true,false]);
    auto.showPopover();
    assertState([true,true]);
    hint.hidePopover();
    assertState([false,false], 'closing the parent hint should close downgraded, nested auto');
    },'Opening an auto popover inside a hint popover is allowed and it downgrades to hint.');
  </script>
</div>

<div>
  <div popover>Auto
    <div popover>Nested Auto</div>
    <div popover=hint>Nested hint</div>
  </div>
  <script>
  test(() => {
    const auto = getPopovers()[0];
    const auto2 = getPopovers()[1];
    const hint = getPopovers()[2];
    auto.showPopover();
    auto2.showPopover();
    assertState([true,true,false]);
    hint.showPopover(); // This should NOT hide auto2, since hint doesn't close unrelated autos.
    assertState([true,true,true]);
    auto.hidePopover(); // Hiding an auto hides its descendant hint stack.
    assertState([false,false,false]);
  },'If you: a) show a popover=auto (call it D), then b) show a descendent popover=hint of D (call it T), then c) hide D, then T should be hidden. (A popover=hint gets hidden when its ancestral popover=auto is hidden)');
  </script>
</div>

<div>
  <div popover>Auto</div>
  <div popover=hint>Non-Nested hint</div>
  <script>
  test(() => {
    const auto = getPopovers()[0];
    const hint = getPopovers()[1];
    auto.showPopover();
    hint.showPopover();
    assertState([true,true]);
    auto.hidePopover();
    assertState([false,true]);
    hint.hidePopover();
    assertState([false,false]);
    },'If you: a) show a popover=auto (call it D), then b) show a non-descendent popover=hint of D (call it T), then c) hide D, then T should stay open. (Non-nested popover=hint does not get hidden when unrelated popover=autos are hidden)');
  </script>
</div>

<div>
  <div popover>Auto
    <div popover>Auto2</div>
    <div popover=hint>Hint</div>
    <div popover>Auto3</div>
  </div>
  <script>
  test(() => {
    const auto1 = getPopovers()[0];
    const auto2 = getPopovers()[1];
    const hint1 = getPopovers()[2];
    const auto3 = getPopovers()[3];
    auto1.showPopover();
    auto2.showPopover();
    hint1.showPopover();
    const events = [];
    auto2.addEventListener("beforetoggle", (e) => { if (e.newState === "closed") events.push("auto"); });
    hint1.addEventListener("beforetoggle", (e) => { if (e.newState === "closed") events.push("hint"); });
    auto3.showPopover();
    assert_array_equals(events, ["hint", "auto"], "Hint should be hidden before Auto");
  },"When showing an auto popover, hint popovers should be closed before auto popovers.");
  </script>
</div>
